1.YTiled增加坐标变换方法：像素坐标转换为引擎坐标 2.修改TiledBean对象层坐标类型为float

+++ YGameFrame/src/ygame/extension/domain/tilemap/YTiledBean.java
@@ -233,12 +233,12 @@ public final class YTiledBean
 	public static class YObject
 	{
 		private YPolygon[] polygon;
-		private int x, y;
+		private float x, y;
 		private String name;
 		private YPolygon[] polyline;
 		private YObject.YProperties properties;
 
-		public int getY()
+		public float getY()
 		{
 			return y;
 		}
@@ -248,7 +248,7 @@ public final class YTiledBean
 			this.y = y;
 		}
 
-		public int getX()
+		public float getX()
 		{
 			return x;
 		}
@@ -327,9 +327,9 @@ public final class YTiledBean
 
 	public static class YPolygon
 	{
-		private int x, y;
+		private float x, y;
 
-		public int getX()
+		public float getX()
 		{
 			return x;
 		}
@@ -339,7 +339,7 @@ public final class YTiledBean
 			this.x = x;
 		}
 
-		public int getY()
+		public float getY()
 		{
 			return y;
 		}
